requirements.txt
Django==1.11.5
olefile==0.44
Pillow==4.2.1
PyMySQL==0.7.11
pytz==2017.2
